# docs/api/openapi-generated.yaml
openapi: 3.0.3
info:
  title: Rate UKMA API
  version: 1.0.0
  description: An API for rating UKMA courses and professors
paths:
...
  /api/v1/students/me/courses/:
    get:
      operationId: students_me_courses_retrieve
      description: List all courses that student is/was enrolled in with information
        about the rating.
      summary: Student's statistics on course rating.
      tags:
      - student
      - courses
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentRatingsLight'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                ValidationError:
                  value:
                    detail: Validation failed
                    status: 400
                    fields:
                      difficulty:
                      - Must be between 1 and 5.
                  summary: Validation error
          description: Bad request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided
                    status: 401
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                Forbidden:
                  value:
                    detail: You do not have permission to perform this action
                    status: 403
          description: Forbidden
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        detail:
          type: string
        status:
          type: integer
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required:
      - detail
      - status
...
    InlineCourseOffering:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        course_id:
          type: string
          readOnly: true
        year:
          type: integer
          readOnly: true
        season:
          type: string
          readOnly: true
        can_rate:
          type: boolean
          readOnly: true
        rated:
          allOf:
          - $ref: '#/components/schemas/InlineRating'
          readOnly: true
          nullable: true
    InlineRating:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        difficulty:
          type: integer
        usefulness:
          type: integer
        comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        is_anonymous:
          type: boolean
      required:
      - difficulty
      - is_anonymous
      - usefulness
...
    StudentRatingsLight:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        offerings:
          type: array
          items:
            $ref: '#/components/schemas/InlineCourseOffering'
          readOnly: true
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"